Fixing and exposing the name finding in a string
Turns out that some browsers don't let you set a function's name
property.  I knew that from before, but then it is still far easier to
diagnose and test the fallback code when I can have any JavaScript
engine hit it.  Thus, now you can pass in a string representation of a
function to this new method and get back the function's name.

Fantastic for testing!